split redux file
