+++ app/javascript/packs/actions/chat.jsx
@@ -1,0 +1,13 @@
+/* Actionsの実装 */
+
+// Action名の定義
+const SEND = 'SEND';
+
+// Action Creators
+export default function send(value) {
+  // Action
+  return {
+    type: SEND,
+    value,
+  };
+}
